* In order to fix partial payments, simplify payments.pm
This commit moves the rather untransparent sub-array of invoices
to a hash with comprehensible key names; this eliminates the need
to suppress critic warnings and clarifies code. Note that this
impacts the output template, because fields now need to be
referenced by key instead of number.

This commit separates (better than before) the data from the
database and the data from the request by accessing each
through their own accessor variable.

This commit simplifies linking invoices from the database
with invoice data from the request by changing the
structure of the form element 'name' attributes.