Use `get` not `get_one` to retrieve request parameters
In lsmb, we've been using `Hash::MultiValue`'s `get_one` method
to retrieve requst parameter values, in the mistaken understanding
that this will return a single value if the request contains
multiple parameters with the wanted key.

This is mistaken. In fact, from the `Hash::MultiValue` documentation
for `get_one`:

> This method croaks if there is no value or multiple
> values associated with the key

We actually want to use the `get` method:

> Returns a single value for the given $key. If there are multiple
> values, the last one (not first one) is returned.

This PR changes uses of `get_one` to `get`. As `get` returns undef
when a given parameter key is not present, rather than croaking,
these calls no longer need to be wrapped in an `eval` block.