Fix Account Heading screen not displaying parent heading
The Account Heading edit screen should display the current parent
heading, but was not doing so because it was using an incorrect
request parameter.

The account heading parent field contains a `NONE` element. The
insertion of this has been isolated so as not to interfere with
the ordinary account creation screen, where a null heading
is invalid.

The legacy null account heading entry has been removed from the
account editing screen, where it is invalid.