Add 'Create Account' 'Create Heading' buttons to Chart of Accounts
As discussed in chat, add buttons to the Chart of Account screen
to add new headings and accounts.

This separates the function of adding headings and accounts, which
were previously separate tabs on the same screen. It groups the
Chart of Account functions together. It allows easy review of
existing accounts and headings before adding more. It removes
clutter from a main accounting menu - adding accounts being only
an occasional administrative function.