Fix Saving Groups clears all user permissions
Commit d8b82c0123badcfe7a3f6294fea43acca592f194 changed the way
groups/permissions are handled. We use a company-specific prefix
for each 'global' role. That commit moved the prefix handling
into the database layer, meaning the perl and template code now
deals only in 'global' role names, without a prefix.

`LedgerSMB::Scripts::contact` was still expecting role names to
be prefixed and was applying a parameter filter to that effect. As
a result, on update, it was as though all roles had been deselected.

This PR fixes that issue, but note that, on form submission, role names
are still prefixed (though no longer company-specific) so that role
parameters can be distinguished from other form parameters.