* Fix #3933: Saving orders deletes unrelated invoice shipping info
The orders should be using the 'oe_id' column, but are in fact
using the 'trans_id' column to store shipping information,
even for deletion. However, the 'trans_id' is used for AR/AP
invoices and thus deletion removes shipping info from invoices...

This commit addresses the orders part of #3899 as well,
although through a very different pattern.