* Make sure the trial balance generator function exists during MC validation
Travis found that I overlooked the fact that our modules have
not been loaded yet on new and migrated (e.g. from SL) schemas
which means that if we want to use functions, we need to copy
them verbatim into the 'temporary namespace' -- so they
automatically get deleted at the end of the session.