* When the input_prefix is set, also prefix the row counter value 'row'
As multiple tables on the same page would be generating
the same row_1, row_2, etc values (which map the row
number to an 'id' value), we need to prefix the name.

As nothing in the code currently uses the prefix,
using the prefix seems like the lowest impact
move to de-duplicate the name attributes.