* MC migration: Work around ar_check and ap_check
The data in the ar/ap tables must abide by the ar/ap_check
constraints which dictate that either currency *and* amount
are NULL or that both are not null.
This means we can't set the currency to a valid value without
setting the amount to a non-NULL value.
Also, the only case where a currency is missing *can* be
the case where the amount is NULL.