Allow coderef to specify select element option list
This commit extends the select element UI template to accept
a coderef to specify the option list. This coderef is called
for each row, with the row fields as an argument. This enables
the option list to be dynamically tailored for each row.

The old behaviour of providing a static arrayref specifying options
common to all rows is still supported and unaffected by this change.