SchemaChecks::JSON - allow callback to specify drop down options
Previously the HTML dropdown UI element was extended to support dropdowns
options to be specified on a per-row basis via a callback. This extends
the SchemaChecks JSON output in the same way.

`options` is now an array reference, with each element corresponding to
a grid row and specifying drop-down options for each field on a row-by-row
basis:

    options => [
      {
        field1 => {
          # Options for Row 1, field1
          value1 => 'Option 1',
          value2 => 'Option 2'
        },
        field2 => {
          # Options for Row1, field2
          value1 => 'Option 1'
        },
      },
      {
        field1 => {
          # Options for Row 2, field1
          value3 => 'Option 3',
          value4 => 'Option 4'
        },
        field2 => {
          # Options for Row2, field2
          value1 => 'Option 1'
      }
    ];