Present custom summary account_link descriptors in UI
This commit extends the Chart of Accounts edit screen to include
custom summary account_link descriptors in the 'Summary Account For'
dropdown.

This dropdown is now populated from the database
`account_link_description` table, rather than being hard-coded (but
required to correspond).

Checks in the perl code to limit the number of summary account_link
descriptors linked to an account have been removed. The UI prevents
this and the database now has a constraint to enforce this. The
perl checks are therefore unnecessary.