* Fix AR/AP account dropdown and 'missing currencies' error
The dropdowns were empty before this commit; apparently a botched
merge-integration when the master branch was updated with smarter
code for dropdown generation.

The various pages for AR/AP transactions/invoices were complaining
on update about missing currencies. As the currencies aren't
actually missing, update the 'update' functions to initialize
$form->{currencies} correctly and stop the complaints.