* Replace inline query for used-credit-limit calculation with stored function
Note that this function changes the definition of the limit to be
in the customer/vendor's preferred currency instead of in the
company's functional currency. For the unsuspecting user, this is
expected to yield the most straightforward outcome as this means
that the limit's used amount will equal the sum of all unpaid
invoice amounts (and unfulfilled orders).

The calculation gets more complex (and harder to understand
) when the customer/vendor has invoices in another currency
than its preferred currency.