Respect tempdir when rendering ODS template
Rendering an ODS document using `OpenOffice::OODoc` requires a
working directory for temporary files.

We use File::Temp to create and destroy this temporary directory,
but it was not being properly configured for the OpenOffice::OODoc
module.

As a result, OpenOffice::OODoc was falling back to its default of
using the current directory. If this was not writeable, template
rendering would fail.

This patch correctly configures the module to use a writeable
temporary directory which is cleaned-up afterwards.