* Fix #2567 and a bug in the now-used code
Instead of doing nested IN constructs, use nested EXISTS,
which should perform *much* better (however, I lack the
database of sufficient size to be able to do representative
measurements)

Note that the commented code didn't check if it needed
to check the locations table for the 'in_country_id'
parameter. Although unlikely to be used by itself,
it *is* an option to do so from the point of view
of the stored procedure, so it's included for completeness
now.