* Fix invoice/transaction failing to save; fix currency reset to default on update
Before this change, invoices and transactions were indefinitely being
recalculated under the assumption that form data had changed since
the last submission. Even when that wasn't the case. This was caused
by a call to invoice_links and/or create_links too early in the
update process.

Note that the change in ir.pl from 'all_vc' to 'create_links' is an
extension of scope: all_vc includes a call to create_links, but does
some extra work, such as initializing the list available currencies.

The call to 'invoice_links' was replaced by a call to Form.pm's
create_links because 'invoice_links' changes the currency back
to the entity's default currency, even when the user changed it
to some other value. 'invoice_links' includes a call to
Form.pm's create_links.