* Validate number of payment links per transaction
Obviously, due to 'moving' payment links from one entry_id to another,
there's the risk of moving payment links between transactions. As such,
we really need to make sure that no transaction ends up with more
payment links than it had or than any transaction has fewer payment
links than it ought to based on the number it had and the number of
lines that were deleted from the transaction.