* Simplify comparison between lines
Using the 'is not distinct from' operator, NULL values get compared as
if they are actual values, meaning that there's no need for specializing
the 'values are NULL' condition.

Similarly, for 'source' and 'memo', consider the NULL and empty string ('')
values to be equal. Doing so allowed 20 more lines to be merged into other
lines (i.e. a reduction of 20 lines) on one of my test databases.