* Sort currency list, default currency first
Note that \@nick_prater found that on his system (with GBP-based books),
the functional currency presented on an invoice was USD (should have
been GBP) where the first line in the 'currency' table was the 'USD' line.

The first row in the \@{$form->{currencies}} array is to be the
default currency. While Form.pm takes care to  set up the value the
way that's required, the other source of currencies was returning
the 'currency' table's natural order. This commit changes that and
makes sure the default currency gets returned first; sorted by
'curr' after that.