* Change type of column exchangerate_default.valid_to to 'date'
When I created the schema, I was under the impression only timestamps
could hold the value 'infinity'. As it turns out, dates can do so as well.
As such, it's not required that valid_to is a timestamp and it's better
to have valid_from and valid_to with equal types.