* Simplify recurring transactions query and screen generation
Now that we have 'MC', there's no need for an exchangerate field in the
query anymore.
Instead of generating an 'ar' module and an 'invoice' indicator, which is
later used to override the 'ar' value with 'is', simply generate the 'is'
module indicator directly in the query, eliminating the need for the
'invoice' column on output.