* Don't inject $form values into SQL statements
Note that this isn't an actual SQL injection problem because the value
was retrieved from the database just a few lines up, but simply the
fact that this requires an inline comment indicates that this pattern
requires brain-attention. Following general patterns frees the brain
power for more complex issues. So: simply following the regular
pattern.