* Remove unused variable ($uid)
The $uid variable was part of a now-removed pattern where INSERTs
would be seeded with a special value based on $uid, to allow finding
back the record. These days, that pattern has been replaced by
manually querying the "next value" of the applicable sequence (in
this case 'oe_id_seq'), just above the variable declaration.