* Orders and quotations fail to save due to employee foreign key
The foreign key on the AR and AP tables points to person.entity_id;
on the oe table, it points to person.id. However, the code in oe.pl
uses the same logic to compose the value for the person_id field
as does the logic in ir.pl/is.pl. Correct the fkey to align with
ir.pl/is.pl.