* Move role grants to Roles.sql
While loading some database to be repaired while helping a user,
the fact that some GRANTs are in Pg-database.sql means they don't
get applied upon reloading the modules -- with resulting in access
problems.
This may be a problem with restores across PG clusters too, when the
user elects not to load the roles file.

Note that the 'chart' view exists in Pg-database, but gets dropped
during "schema change updates", meaning the grant on it can't be
executed in Roles.sql: by the time it gets there, the view doesn't
exist anymore.