* Fixes #4016: tax doesn't post on AA transaction
Note that in previous code, the (non-fx) amount was made up
of the difference between the fx amount and the total amount
to be posted in the base currency.
In the code that posts AA transactions, this paradigm hadn't
been completely replaced with the new one which posts full
amounts both in the amount_bc as well as in the amount_tc
columns.
Combined with the fact that the "fx adjustment" is guaranteed
to be zero (0) for any base-currency transaction, the adjustment
amount is guaranteed to be zero and hence the criterion for the
posting to be made when the (difference) amount is non-zero,
always fails the test and thereby always fails to post taxes.