* Don't test modules being installed (and make sure to try and use XS code for JSON)
Note that the wire protocol for Selenium::Remote::Driver is
entirely JSON. So, when in code coverage mode, the performance
of the JSON parser module will be greatly implicated, even if
the module is out of scope (or at least, that's the assumption
underlying this commit).
Lets see what happens when we switch to XS (and thus outside
the realm of code coverage measurementy).