* Ask PGObject::Util::DBAdmin for runnable 'psql'
Instead of checking ourselves, we should ask our library to tell us
whether it can successfully do what it's supposed to do. If we do it
ourselves, we have a layering violation *or* we need to replicate the
logic the library uses to assure execution environment. Both are highly
undesirable.