Fix 'pricelist' link on contacts screen
When clicking the link, the server response is said to be faulty,
which is caused by the fact that the HTML being returned includes
2 'dynatable' results, both of which contain a price matrix; one for
the customer and one for the price group the customer is part of.

Unfortunately, the 2 dynatables use the same 'id' attribute values.
This changes adds the prefix to be used for the 'name' attribute
to the 'id' attribute values as well.