Use Starlet in order to allow faster responses from the application server
Using Starlet, we have 2 workers, which means that the http server hosting
the application will be more responsive to requests being issued in quick
succession.
This both speeds up testing as well as reduces the chances of requests being
sent, but not timely handled because they get stuck in limbo (or in the
socket queue).