Add back checkmarks to mark account AR/AP 'tax applicable'
These checkmarks were lost on the 1.2->1.3 rewrite with the effect
that accounts without this checkmark will show on the AR/AP
screen, but associated amounts will not show up when viewing
(=loading from the database) a posted transaction.

Also sort the 'Tax' checkmark for each of the checkmark rows into
the same column.