* Fix errors when restoring backups
pg_dump includes the ~*~ custom operator later in the dump stream
than its use in the employee_search function. This throws an ERROR
and fails to create the function.

Because the operator is used only in 1 function, lets not make things
complicated and simply remove it in favor of expanding the wrapper
it really is.