Fix tax checkmarks sometimes being automatically selected
Due to an error in the regex (the $-sign is expanded to an empty
string), only prefix matching happened before this change. In a
specific database, one tax account was numbered 21 while another
was numbered 215. Checking the account with id 21 resulted in both
the 21 and 215 account being checked in the UI.