Print checks in numbered order
This is a regression in 1.7 due to the 'parameter parsing' step
introduced after 1.6 which maps the 'flat' argument hash into a
hierarchical structure which processes more easily.

Due to hash randomization the order of the payments (and thus
their source numbering) is non-deterministic. Sort the source
numbers (=check numbers) to make sure they are processed in
source number order.