Skip COGS calculation when posting a service
Before this change, on reversal (negative quantity), code gets executed
which assumes an inventory account to be available. Obviously, a service
isn't kept in inventory, meaning that there's no inventory account and
the code fails hard.

Exiting early when there's no inventory account prevents this scenario
(and all the others which require an inventory account).