Fix first reconciliation failing to be displayed
Under some circumstances, the first reconciliation for an account fails
to be correctly displayed with an error from the BigFloat module saying
that 'is_zero' can't be called on an undefined value.

The cause seems to be that the first reconciliation on an account
doesn't have a defined 'our_total' -- as there's no prior recon.

When there's no defined value, default to zero (0).