Fix bug preventing first transaction on account appearing in recon report
The `reconciliation__pending_transactions` function tries to be smart
and efficient by excluding all transaction preceeding the first one
which was posted on the account to be reconciled. However, in the
process, it also excluded the first transaction itself.
Moreover, the exclusion of the preceeding transactions is largely
irrelevant, because over the lifespan of the database, the number of
*succeeding* transactions will be an order of magnitude bigger than
the excluded ones. Yet the succeeding transactions will mostly have
been reconciled, so be considered spurriously too.