Rollback database state in case of an execution error
The exact same has been implemented on 'new code' but the definition of
"an error condition" isn't as clear in old code as it is in "new code".

The existing code *always* commits the latest state, which means that if
the error doesn't come from database code (in which case a commit will
fail, essentially rolling back), the database state is committed while
the actual processing didn't run to completion.

After this change, at least all identified error cases, the (partial)
database state will be discarded, even if the source of the error isn't
in the database.