Reduce the dependency on LedgerSMB::App_State::DBH
This commit moves `set_datestyle` to Form.pm because Form.pm
is the only user *and* it's deprecated old-code. This way the
routine will be automatically cleaned up when we remove Form.pm.

Instead of jumping through hoops trying to estimate whether
$LedgerSMB::App_State::Company_Config might have a value,
check the value directly. At least this reduces the dependency
on LedgerSMB::App_State::DBH...