Use `LSMB_ALWAYS_MONEY` to convey the number of digits to round to
Instead of signalling specific rounding is requested, only to
request the specific rounding at a much deeper level in the
call-stack (a place where no access is available to the database
handle required to retrieve said information).

This change has the effect that instead of the database handle being
marshalled through the call-stack, the number of digits is.

Additionally, instead of retrieving the number of formatting digits on
every formatted number, this change retrieves the number only once,
saving (x-1) queries -- with x being the number of items being formatted.