Add infrastructure to render reports either as document or UI element
At the moment report rendering is a mix of document rendering and
UI element rendering: it uses templates stored in UI/ (a pattern
for UI elements) yet uses the LedgerSMB::Template renderer (a
pattern for documents).

This commit factors out the tie to both the template source as
well as the document template renderer.