Remove the 'dev manual'
The document is just a skeleton, apart from the chapters on the
"workflow scripts" and the "model classes", both of which seem to
have been generated from an old version of the sources.

All in all conclusion is that the document isn't relevant to the
development process and source tree as it stands.