Follow-up to fixing #2587 (9faef0ad2a61d9cd1c771bf7a20a3979c747366c)
Fixes the fact that all buttons show, instead of only the applicable ones.

Note that this commit causes only 1 button (Update) to show on the initial
screen, which is completely correct: without a vendor having been selected,
the data on the screen can't save and therefor the button shouldn't show.

(Also, the system can't calculate whether or not the transaction will be
posted in an "open" period, due to the missing 'transdate' which the
browser must provide.)