Implement new template management UI
The UI implements a much more dynamic interaction than
the exsiting UI. Also, the new UI provides feedback about the
(non)existence of templates in the database directly; the old
UI would provide erratic error messages.

This solves the bug that there was no way to create a template
for a language+format which wasn't already in the database (the
old UI would generate errors in that case).

Enormously reduces the size of the menu.