Remove identity 'escaping' and HTML escaping from ODS, XLSX, CSV and TXT
Note that Template Toolkit supports escaping out of the box for HTML/XML;
all we need to do is use it upon insertion into the template.

Note that this makes processing of these formats more efficient, because
the supplied variables (usually the $form/$request hash) doesn't need to
be completely traversed for every document; instead, only inserted values
need to be escaped (for ODS+XLSX) -- using TT standard functionality.