Add missing payments to the payments table
As part of the effort to consolidate all payment information into a
single place, this commit copies all existing payment information
into the payments table, using the definition of 'payment' as
implemented in the 'payment__search' stored function.

Note that this commit creates payment records for payments entered
into the invoice and transaction screens (which lack both a payment
record as well as a link to a voucher) as well as use existing
payment/receipt voucher information to help identify payments.

Closes #4140