Fix warning from xt/41-coaload.t test
When running xt/41-coaload.t as part of our development test suite,
a warning was produced:

`Use of uninitialized value $dir in concatenation (.) or string at xt/41-coaload.t line 32`

The root cause of this is a regex used to parse the input sql file names
to generate temporary test database names.

This patch changes the way temporary test database names are derived from
the file names, such that a warning is no longer generated.