Fix SQL error migrating databases with unapproved reconciliation reports
The 1.8 migration process requires that unapproved reconciliation reports
are deleted.

This operation yielded an error message of the form:

    Failed to remove unapproved report: ERROR: syntax error at or near "'{"272"}'"
    LINE 1: DELETE FROM cr_report_line WHERE report_id IN '{"272"}'
    ^ at /srv/ledgersmb/sql/changes/1.8/explicit-reconciliation.sql.checks.pl line 46.

This is because the code tries to pass an array reference of ids as a DBI
placeholder value, which is explicitly prohibited by DBI. From the documentation:

  ...placeholders can only represent single scalar values. For example,
  the following statement won't work as expected for more than one value:
    "SELECT name, age FROM people WHERE name IN (?)"    # wrong
    "SELECT name, age FROM people WHERE name IN (?,?)"  # two names

The code has been refactored to make a separate database call for each
report being deleted. While this is more round-trips to the database, it
maintains proper escaping through use of placeholders.

This fixes the error.