Don't migrate zero or null exchange rates as defaults
In lsmb 1.6 and earlier, exchange rates were stored on a per-day
basis, always as a buy and sell rate pair.

In legacy databases, often only one of the buy or sell fields is
populated for a given day. In this situation, the unpopulated field
may be set to `0` or `NULL`.

For lsmb 1.7, we migrate these exchange rates into an `exchangerate_default`
table. This patch ensures that we do not migrate the unpopulated `0` or
`NULL` values.

Note that the `exchangerate_default` table is only used to provide default
exchange rates for convenience when entering a new transaction. It has no
impact on accounting data.