Add NOT NULL constraint to exchange rate defaults table
This patch ensures that the exchangerate_default.rate field cannot
be `NULL` as that is an invalid value, which causes errors in other
parts of the code that tries to use it (such as the Edit rates screen).

During the upgrade, migration checks are used to identify any existing
NULL exchange rate defaults and the upgrader is asked to accept their
deletion.