Change to encrypted cookie storing a session object
As opposed to the cookie we used to use which stored - in plain text -
the session ID, name of the database logged into and the session token.

This change is a step to get away from Basic Auth where we can't control
the browser popup nor the expiration of the credentials. Also, in the
modern world of web technology, basic auth is a very limited form of auth
where the web is currently adopting much more complex auth methods such
as OAuth, WebAuth and FIDO2. Ultimately, we want to tap into those
methods!