Move the main app to encrypted cookie authentication
This commit eliminates the authentication popups browsers show in
case of expired sessions or incorrect configurations. Additionally,
this commit reduces the complexity of the authentication code path.

Also note that this change removes 'logout_js' which was used to
generate a 401 authentication meant to clear the credentials cached
in the browser. This is no longer necessary as the cookie will be
removed due to an 'Expires' setting in the past.

This commit introduces a new configuration key for ledgersmb.conf:
`cookie_secret`: the secret to be used to encrypt the cookie. When
none is provided, a secret is automatically generated.