Remove and replace access checks in Form::new()
Due to the way request dispatched is implemented, the check whether
whitelisted scripts are being called is ineffective and can be skipped:
the same check happens as part of the outer request dispatch.

Checking the name of the action for invalid characters is replaced
by a check if the target action has been defined.