Consolidate common BDD steps
In developing bdd tests for various parts of LedgerSMB, there are various
steps that are common to a number of screens. This stems from our re-use
of UI components, such as dynatable for displaying tabular information.

This PR is a first step in aiming to reduce the copy/paste code used to
interact with such common components.