Use SessionStorage middleware with old code
In order to be able to do so, it should be possible to marshall deserialized
session state into old code. This commit makes that happen. Unfortunately,
the inner scope forks, meaning that the session state doesn't come back
out... (and the inner scope must generate the cookie itself)

Additionally fix the fact that the store variable is now located in
SessionStorage and no longer in AuthenticateSession.