BDD test framework - use class name to find dynatable element
Previously when interacting with dynatable page elements we searched
for the first table on the page with data rows. This generally works
for our reports, but breaks if the report returns no data rows.

This change adds a qualifier to ensure that only table elements
having the `dynatable` class are considered as report data rows.