Take $env as an argument for auth callbacks (instead of closing over it)
By closing over $env a circular reference is created by storing a
reference to the closure in the $env hash. This circular reference
means the hash remains allocated after the request returns.

By requiring the environment to be passed as an argument,
we're not closing over the environment, preventing the circular
reference.