Use uri not url escaping for dynatable row_id href
The Template Toolkit `uri` filter "corectly encodes all reserved characters,
including &, @, /, ;, :, =, +, ? and $" which is what we need for a query
string parameter.

Previously we were escaping the whole url, including query parameters,
using the Template Toolking `url` filter, which does not encode these
reserved characters, meaning that they were not being correctly encoded.

This patch eascapes the query parameter separately to the base url so that
escaping is properly applied.