1.7 Fix recipt totals wrong with number format
When a number format is in use (for example formatting an amount
with thousands and decimal separators as 1,234.56), the totals
on the Cash->Receipts screen were calculated incorrectly.

This is because the numbers were not being converted from the
display/input format before calculation. Logs were displaying
messages of the following form:

   Argument "1,234.56" isn't numeric in subroutine entry at /srv/ledgersmb/lib/LedgerSMB/Scripts/payment.pm line 1178.

This patch resolves this issue by converting such numbers using
`LedgerSMB::PGNumber->from_input` before calculating the total.

A similar patch is required for master, but as source code has
diverged, this patch cannot be directly applied.