Change coding for readability (other pgTAP primitives) and headings
Note that this commit aims to improve readability of the test code
and only that. It does it in 2 ways: better separate the parts of
the tests by introducing better visually separating headings as
well as the use of more diverse pgTAP primitives, which result in
shorter, more concise coding.