Test (and fix) discount calculation on payment terms
Before this commit, payment discounts would be calculated based on 'today'
instead of on the date of the payment. This commit uses the 2-argument form
of Pg's `age()` function to explicitly specify the base.

Note that using EXTRACT(DAYS) is the wrong way to check against the payment
terms, because extracting DAYS from the interval '1 month 15 days' returns
*15*, not 45. Instead, use interval mathematics directly.

Fixes #1494