Remove the source dropdown from the payment screen
Closes #2320

As documented, it is, it's just a prefix to the actual value stored in the
'source' field in the database, thus cluttering the source field for those
who want to use its clean value.

In the mail discussion following https://archive.ledgersmb.org/ledger-smb-devel/msg06519.html,
Chris Travers documented that it was added for POS purposes *and* that the
POS functionality has been removed; orphaning this functionality.