Make absolutely sure the database handle gets cleaned up
Note that in case of an exception, the result callback may not be
executed (the entire stack frame will be skipped). To be certain the
database handle gets cleaned up, register a scope handler to do
cleaning up as soon as the environment gets discarded (at the end
of the request...).