Force CAMT053 import to interpret content as string, trap fatal errors
When uploading a bank reconciliation file, we slurp its contents into
a string before passing on to the various parsers. Previously, if
the content did not include '<' or '>' characters, the CAMT053
importer would treat the string as a filename and attempt to read
from the filesystem. We now explicitly treat the input as a string
of (possible) XML.

Additionally, we expect the CAMT053 parser to gracefully reject
invalid data. To perform the detection, we attempt to parse the
input as XML. The parser will die if the input is not valid xml. We
therefore now catch parser errors in a try{} block so that they are
not fatal.