Replace XML::Simple with XML::LibXML
This PR replaces use of XML::Simple with XML::LibXML.

XML::Simple is documented as `PLEASE DO NOT USE THIS MODULE IN NEW CODE`.

This affects the ISO20022 CAMT053 bank statement file import, which
has been refactored to use the new XML module.

This change does not introduce a new dependency to lsmb. Rather
it removes a dependency, as XML::LibXML was already a dependency
of the XML::Simple module previously in use.